Move flush test case to separate folder
